requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langgraph==0.2.14
langgraph-checkpoint==1.0.12
langsmith==0.1.147
langchain-google-genai==1.0.10
google-generativeai==0.7.2
langchain-experimental==0.0.62
SQLAlchemy==2.0.35
mysql-connector-python==9.0.0
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.2